+++ crates/editor/src/element.rs
@@ -278,6 +278,7 @@ impl EditorElement {
         register_action(editor, window, Editor::transpose);
         register_action(editor, window, Editor::rewrap);
         register_action(editor, window, Editor::cut);
+        register_action(editor, window, Editor::kill_line);
         register_action(editor, window, Editor::kill_ring_cut);
         register_action(editor, window, Editor::kill_ring_yank);
         register_action(editor, window, Editor::copy);